weather app/backend/requirements.txt
fastapi 
uvicorn 
pydantic 
psycopg2 
python-dotenv